Validate length of Cms::Page slug and full_path
Currently databases are just chopping off the tail end of any slug or full_path that is longer than the field lengths. With this change a helpful error message will be displayed to the user.

We also added a view only field of full_path in the admin edit page, to show this error message.

We also added some handy styles for disabled inputs.


Former-commit-id: 8196b41bec058d6e3512dd2148f626bf77881add